+++ src/main/java/me/RockinChaos/cloudsync/bungee/utils/ServerUtils.java
@@ -15,9 +15,9 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package bungee.me.RockinChaos.cloudsync.utils;
+package src.main.java.me.RockinChaos.cloudsync.bungee.utils;
 
-import bungee.me.RockinChaos.cloudsync.utils.api.SnapAPI;
+import src.main.java.me.RockinChaos.cloudsync.bungee.utils.api.SnapAPI;
 
 public class ServerUtils {
