+++ src/main/java/me/RockinChaos/cloudsync/bungee/utils/api/SnapAPI.java
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package bungee.me.RockinChaos.cloudsync.utils.api;
+package src.main.java.me.RockinChaos.cloudsync.bungee.utils.api;
 
 import java.util.logging.Logger;
